[UI] adding refresh after adding new button

+++ code/src/Entity/Product.php
@@ -85,15 +85,15 @@ class Product
     }
 
     /**
-     * @return int
+     * @return float
      */
-    public function getRating(): int
+    public function getRating(): float
     {
         return $this->rating;
     }
 
     /**
-     * @param int $rating
+     * @param float $rating
      */
     public function setRating(float $rating): void
     {
